web: ak-checkbox-group for short, static, multi-select events
Implements a checkbox groups web component, wholly independent of the API
(although it does implement the 'data-ak-control' protocol, including the
`json()` method that makes it easier to send the data to the Form handler).  The
controller works much like multi-select: `value` returns an array of strings,
the `name` attribute associated with whatever it is you're asking about.

The `required` property only works if you give the whole item a name, as if it
were an input.  Otherwise, it does nothing.

Giving it a `name` also activates the browser standard `formAssociated`
protocol; it works just fine for ordinary HTML forms, and presents to that
protocol the `FormValue` type, so any form using it will automagically convert
it into the CGI (Common Gateway Interface) format of, to use the example from
Storybook:

```
ak-test-checkgroup-input=funky&ak-test-checkgroup-input=invalid
```

Note that the classic CGI format is not automatically key/value; keys can appear
multiple times, and indicate that the value is an array of strings.  Most modern
appservers understand this format. Some do not.

There's a full and complete JSDOC-like comment documenting the component.  I
have even provided CSSPart sections for everything: the wrapper, each line, the
input and its associated label.  The brave or foolhardy can mangle the CSS to
their hearts' content without having to know a thing about Patternfly.